% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{bootCI}
\alias{bootCI}
\title{Bootstrap CIs}
\usage{
bootCI(d1, d0, tau = NULL, dist1, dist0, B = 2000, sig)
}
\arguments{
\item{d1}{Target data.frame containing time and status.}

\item{d0}{Reference data.frame containing time and status.}

\item{tau}{Optional truncation times for calculating restricted mean survival
time.}

\item{dist1}{String, target distribution.}

\item{dist0}{String, reference distribution.}

\item{B}{Number of resamples.}

\item{sig}{Significance level, for CIs.}
}
\value{
List containing data.frames with the bootstrap CIs for the location
  and RMST estimates.
}
\description{
Generates bootstrap confidence intervals for location and RMST estimates.
}
