% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{checkArm}
\alias{checkArm}
\title{Check Arm}
\usage{
checkArm(arm, n)
}
\arguments{
\item{arm}{0/1, treatment arm.}

\item{n}{Integer, sample size.}
}
\value{
Logical indication of whether arm was properly formatted.
}
\description{
Check whether treatment arm is properly formatted.
}
