% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{checkDist}
\alias{checkDist}
\title{Check Distribution}
\usage{
checkDist(dist)
}
\arguments{
\item{dist}{String, distribution name.}
}
\value{
Logical indication of whether the distribution was available.
}
\description{
Check whether the distribution selected is available.
}
