% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{CheckInfoPD}
\alias{CheckInfoPD}
\title{Check Information Matrix Positive Definiteness}
\usage{
CheckInfoPD(inv_info)
}
\arguments{
\item{inv_info}{Inverse of the observed information matrix.}
}
\value{
Invisible \code{TRUE} if check passes.
}
\description{
Stops with an error if the inverse information matrix has any negative
diagonal element (indicating the information matrix is not positive definite).
}
\keyword{internal}
