% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{CheckInit}
\alias{CheckInit}
\title{Check Initialization}
\usage{
CheckInit(dist, init)
}
\arguments{
\item{dist}{String, distribution name.}

\item{init}{List of named parameters.}
}
\value{
None.
}
\description{
Check whether the initialization is valid.
}
