% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{CheckStatus}
\alias{CheckStatus}
\title{Status Check}
\usage{
CheckStatus(status)
}
\arguments{
\item{status}{0/1 status indicator.}
}
\value{
None.
}
\description{
Function to ensure the status indicator is properly formatted
}
