% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{checkTheta}
\alias{checkTheta}
\title{Check Theta}
\usage{
checkTheta(dist, theta)
}
\arguments{
\item{dist}{String, distribution.}

\item{theta}{Numeric, parameter vector.}
}
\value{
Logical indication of whether the appropriate number of parameters
  was provided.
}
\description{
Function to check the appropriate number of parameters are supplied for the
selected distribution.
}
