% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{methods}
\name{show,contrast-method}
\alias{show,contrast-method}
\title{Show Method for a Contrast of Survival Distributions.}
\usage{
\S4method{show}{contrast}(object)
}
\arguments{
\item{object}{An object of class \code{contrast}.}
}
\description{
Show Method for a Contrast of Survival Distributions.
}
