% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{ContrastRMSTs}
\alias{ContrastRMSTs}
\title{Contrast RMSTs}
\usage{
ContrastRMSTs(fit1, fit0, sig = 0.05)
}
\arguments{
\item{fit1}{Fitted parametric survival distribution for arm 1.}

\item{fit0}{Fitted parametric survival distribution for arm 0.}

\item{sig}{Significance level, for}
}
\value{
Data.frame contrasting the difference and ratio of RMSTs
  at each time point.
}
\description{
Compare the restricted mean survival times of the fitted distributions
for two treatment arms.
}
