% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{D1aLogIncGamma}
\alias{D1aLogIncGamma}
\title{First Logarithmic Derivative in Alpha}
\usage{
D1aLogIncGamma(a, x, h = 1e-04, method = "fd")
}
\arguments{
\item{a}{Shape}

\item{x}{Lower limit of integration}

\item{h}{Step size}

\item{method}{Select "fd" for finite difference, "ni" for numeric
integration.}
}
\description{
Evaluates the first derivative of the log incomplete gamma function
with respect to the shape parameter.
}
