% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{D1lIncGamma}
\alias{D1lIncGamma}
\title{First Derivative in Lambda}
\usage{
D1lIncGamma(a, l, u)
}
\arguments{
\item{a}{Shape.}

\item{l}{Rate.}

\item{u}{Time.}
}
\description{
Evaluates the first derivative of the incomplete gamma function
with respect to the rate parameter.
}
