% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{D1lLogIncGamma}
\alias{D1lLogIncGamma}
\title{First Logarithmic Derivative in Lambda}
\usage{
D1lLogIncGamma(a, l, u)
}
\arguments{
\item{a}{Shape.}

\item{l}{Rate.}

\item{u}{Time.}
}
\description{
Evaluates the first derivative of the log incomplete gamma function
with respect to the rate parameter.
}
