% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{D2aIncGamma}
\alias{D2aIncGamma}
\title{Second Derivative in Alpha}
\usage{
D2aIncGamma(a, x, h = 1e-04, method = "ni")
}
\arguments{
\item{a}{Shape}

\item{x}{Lower limit of integration}

\item{h}{Step size}

\item{method}{Select "fd" for finite difference, "ni" for numeric
integration.}
}
\description{
Evaluates the second derivative of the incomplete gamma function
with respect to the shape parameter.
}
