% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{D2lIncGamma}
\alias{D2lIncGamma}
\title{Second Derivative in Lambda}
\usage{
D2lIncGamma(a, l, u)
}
\arguments{
\item{a}{Shape.}

\item{l}{Rate.}

\item{u}{Time.}
}
\description{
Evaluates the second derivative of the incomplete gamma function
with respect to the rate parameter.
}
