% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{D2lLogIncGamma}
\alias{D2lLogIncGamma}
\title{Second Logarithmic Derivative in Lambda}
\usage{
D2lLogIncGamma(a, l, u)
}
\arguments{
\item{a}{Shape.}

\item{l}{Rate.}

\item{u}{Time.}
}
\description{
Evaluates the second derivative of the log incomplete gamma function
with respect to the rate parameter.
}
