% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{D2mixIncGamma}
\alias{D2mixIncGamma}
\title{Mixed Derivative}
\usage{
D2mixIncGamma(a, l, u)
}
\arguments{
\item{a}{Shape.}

\item{l}{Rate.}

\item{u}{Time.}
}
\description{
Evaluates the mixed partial of the incomplete gamma
function with respect to the shape and rate parameters.
}
