% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{D2mixLogIncGamma}
\alias{D2mixLogIncGamma}
\title{Mixed Logarithmic Derivative}
\usage{
D2mixLogIncGamma(a, l, u, h = 1e-04)
}
\arguments{
\item{a}{Shape.}

\item{l}{Rate.}

\item{u}{Time.}

\item{h}{Step size.}
}
\description{
Evaluates the mixed partial of the log incomplete gamma
function with respect to the shape and rate parameters.
}
