% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{DefaultParam}
\alias{DefaultParam}
\title{Set Default Parameters}
\usage{
DefaultParam(dist)
}
\arguments{
\item{dist}{String, distribution name.}
}
\value{
Numeric parameter ist.
}
\description{
Function to select default parameter values for each distribution.
}
