% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormSurv.R
\name{dLogNormSurv}
\alias{dLogNormSurv}
\title{Logarithmic Derivatives of the Normal Surviavl Function}
\usage{
dLogNormSurv(s, Order = 1)
}
\arguments{
\item{s}{Value of \eqn{s} at which to evaluate.}

\item{Order}{Order of the derivative, select from among 1 and 2.}
}
\description{
#' Evaluates logarithimic derivatives of the normal survival, defined as:
\deqn{\ln\Phi(s)=\ln\int_{s}^{\infty}\frac{e^{-(u^2)/2}}{\sqrt{2\pi}}du}
}
