% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{ExtractObsEst}
\alias{ExtractObsEst}
\title{Extract Observed Estimates}
\usage{
ExtractObsEst(fit1, fit0)
}
\arguments{
\item{fit1}{Fitted parametric survival distribution for arm 1.}

\item{fit0}{Fitted parametric survival distribution for arm 0.}
}
\value{
Numeric vector.
}
\description{
Helper function for permutation inference.
}
