% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{fit-class}
\alias{fit-class}
\title{Model Fit}
\description{
Defines the object class returned by fitting functions.
}
\section{Slots}{

\describe{
\item{\code{Distribution}}{Fitted distribution.}

\item{\code{Parameters}}{Parameters.}

\item{\code{Information}}{Information components.}

\item{\code{Outcome}}{Properties of fitted distribution.}
}}

