% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{methods}
\name{show,fit-method}
\alias{show,fit-method}
\title{Show for Fitted Survival Models}
\usage{
\S4method{show}{fit}(object)
}
\arguments{
\item{object}{A \code{fit} object.}
}
\description{
Show for Fitted Survival Models
}
