% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{fit.Gamma.Complete}
\alias{fit.Gamma.Complete}
\title{Gamma Distribution Parameter Estimation without Censoring}
\usage{
fit.Gamma.Complete(time, eps = 1e-06)
}
\arguments{
\item{time}{Observation times.}

\item{eps}{Tolerance for Newton-Raphson iterations.}
}
\description{
Estimates parameters for gamma event times not subject to censoring.
}
