% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{fit.Gamma}
\alias{fit.Gamma}
\title{Gamma Parameter Estimation}
\usage{
fit.Gamma(time, status, alpha = 0.05, eps = 1e-06, maxit = 10)
}
\arguments{
\item{time}{Observation times.}

\item{status}{Status indicator, coded as 1 if an event was observed, 0 if censored.}

\item{alpha}{Significance level, for CIs.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}
}
\value{
An object of class \code{fit} containing the following:
\describe{
 \item{Parameters}{The estimated shape \eqn{\alpha} and rate \eqn{\lambda}.}
 \item{Information}{The observed information matrix.}
 \item{Outcome}{The fitted mean, median, and variance.}
}
}
