% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGamma.R
\name{fit.GenGamma.Complete}
\alias{fit.GenGamma.Complete}
\title{Initialization for Generalized Gamma}
\usage{
fit.GenGamma.Complete(time, bL, bU)
}
\arguments{
\item{time}{Observed event times.}

\item{bL}{Lower limit on possible values for beta.}

\item{bU}{Upper limit on possible values for beta.}
}
\value{
3x1 numeric vector of estimated parameters, \eqn{\alpha},
  \eqn{\beta}, \eqn{\lambda}.
}
\description{
Initializes the parameters for the generalized gamma distribution via maximum
likelihood.
}
