% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{FitGammaComplete}
\alias{FitGammaComplete}
\title{Gamma Parameter Estimation without Censoring}
\usage{
FitGammaComplete(data, eps = 1e-06)
}
\arguments{
\item{data}{Data.frame.}

\item{eps}{Tolerance for Newton-Raphson iterations.}
}
\value{
Numeric vector containing the estimated shape and rate parameters.
}
\description{
Paramter estimation for gamma event times without censoring.
}
