% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGamma.R
\name{FitGenGammaComplete}
\alias{FitGenGammaComplete}
\title{Generalized Gamma Parameter Estimation without Censoring}
\usage{
FitGenGammaComplete(data, beta_lower = 0.1, beta_upper = 10)
}
\arguments{
\item{data}{Data.frame.}

\item{beta_lower}{Lower limit on possible values for beta.}

\item{beta_upper}{Upper limit on possible values for beta.}
}
\value{
Numeric vector containing the estimated shape and rate parameters.
}
\description{
Paramter estimation for generalized gamma event times without censoring.
}
