% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{FitLogNormComplete}
\alias{FitLogNormComplete}
\title{Log-Normal Parameter Estimation without Censoring}
\usage{
FitLogNormComplete(data)
}
\arguments{
\item{data}{Data.frame.}
}
\value{
Numeric vector containing the estimate location and scale parameters.
}
\description{
Log-Normal Parameter Estimation without Censoring
}
