% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{GammaInfo}
\alias{GammaInfo}
\title{Gamma Observed Information}
\usage{
GammaInfo(data, shape, rate)
}
\arguments{
\item{data}{Data.frame.}

\item{shape}{Shape parameter \eqn{\alpha}.}

\item{rate}{Rate parameter \eqn{\lambda}.}
}
\value{
Numeric information matrix.
}
\description{
Observed information for gamme event times without censoring.
}
