% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{GammaScore}
\alias{GammaScore}
\title{Gamma Profile Score for Shape}
\usage{
GammaScore(data, shape)
}
\arguments{
\item{data}{Data.frame.}

\item{shape}{Shape parameter.}
}
\value{
Numeric score.
}
\description{
Profile score equation for gamma event times without censoring.
}
