% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGamma.R
\name{GenGammaObsInfo}
\alias{GenGammaObsInfo}
\title{Generalized Gamma Observed Information}
\usage{
GenGammaObsInfo(data, alpha, beta, lambda)
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{First shape parameter.}

\item{beta}{Second shape parameter.}

\item{lambda}{Rate parameter.}
}
\value{
Numeric observed information matrix.
}
\description{
Observed information for the generalized gamma log likelihood in the absence
of censoring.
}
