% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGamma.R
\name{GenGammaProfileLogLik}
\alias{GenGammaProfileLogLik}
\title{Generalized Gamma Profile Log Likelihood}
\usage{
GenGammaProfileLogLik(data, beta)
}
\arguments{
\item{data}{Data.frame.}

\item{beta}{Second shape parameter.}
}
\value{
Numeric profile log likelihood.
}
\description{
Profile log likelihood of the generalized gamma distribution as a function
of the second shape parameter \eqn{\beta}.
}
