% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGamma.R
\name{GenGammaRate}
\alias{GenGammaRate}
\title{Generalized Gamma Rate MLE}
\usage{
GenGammaRate(data, alpha, beta)
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{First shape parameter.}

\item{beta}{Second shape parameter.}
}
\value{
Numeric MLE of the rate \eqn{\lambda}.
}
\description{
Profile MLE of the generalized gamma rate given the shape parameters.
}
