% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGamma.R
\name{GenGammaScore}
\alias{GenGammaScore}
\title{Generalized Gamma Score Equation}
\usage{
GenGammaScore(data, alpha, beta, lambda)
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{First shape parameter.}

\item{beta}{Second shape parameter.}

\item{lambda}{Rate parameter.}
}
\value{
Numeric score vector.
}
\description{
Score equation for the generalized gamma log likelihood in the absence
of censoring.
}
