% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGamma.R
\name{GenGammaShape}
\alias{GenGammaShape}
\title{Generalized Gamma Shape MLE}
\usage{
GenGammaShape(data, beta)
}
\arguments{
\item{data}{Data.frame.}

\item{beta}{Second shape parameter.}
}
\value{
Numeric MLE of the rate \eqn{\alpha}.
}
\description{
Profile MLE of the first shape parameter \eqn{\alpha} of the generalized gamma
given the second shape parameter \eqn{\beta}.
}
