% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGamma.R
\name{init.GenGamma}
\alias{init.GenGamma}
\title{Initialization for Generalized Gamma}
\usage{
init.GenGamma(tobs, L = 0.01, U = 10)
}
\arguments{
\item{tobs}{Observed event times.}

\item{L}{Lower bound on \eqn{\alpha}.}

\item{U}{Upper bound on \eqn{\alpha}.}
}
\value{
List containing the estimated shape and rate parameters on log scale.
}
\description{
Initializes the parameters for the generalized gamma distribution using
method of moments.
}
