% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{LogNormInfo}
\alias{LogNormInfo}
\title{Log-Normal Observed Information}
\usage{
LogNormInfo(data, loc, scale, log_scale = FALSE)
}
\arguments{
\item{data}{Data.frame.}

\item{loc}{Location parameter.}

\item{scale}{Scale parameter.}

\item{log_scale}{Is the scale parameter logged?}
}
\value{
Numeric score.
}
\description{
Observed information for log-normal event times without censoring.
}
