% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{LogNormScore}
\alias{LogNormScore}
\title{Log-Normal Score Equation}
\usage{
LogNormScore(data, loc, scale)
}
\arguments{
\item{data}{Data.frame.}

\item{loc}{Location parameter.}

\item{scale}{Scale parameter.}
}
\value{
Numeric score.
}
\description{
Score equation for log-normal event times without censoring.
}
