% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewtonRaphson.R
\name{NewtonRaphson}
\alias{NewtonRaphson}
\title{Newton Raphson Estimation}
\usage{
NewtonRaphson(init, obj, eps = 1e-06, maxit = 10, report = FALSE)
}
\arguments{
\item{init}{Initial value.}

\item{obj}{Objective function.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}

\item{report}{Report fitting progress?}
}
\value{
Numeric parameter estimate.
}
\description{
Newton Raphson Estimation
}
