% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewtonRaphson.R
\name{NRUpdate}
\alias{NRUpdate}
\title{Newton Raphson Update Iteration}
\usage{
NRUpdate(obj, state)
}
\arguments{
\item{obj}{Objective function.}

\item{state}{List containing the parameter vector `theta`.}
}
\value{
List containing the updated parameter vector `theta` and the
  objective increment `delta`.
}
\description{
Newton Raphson Update Iteration
}
