% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.fit}
\alias{print.fit}
\title{Print Method for Fitted Survival Distributions}
\usage{
\method{print}{fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fit}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{fit}.
}
