% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{qLogLogistic}
\alias{qLogLogistic}
\title{Quantile Function for the Log-Logistic Distribution}
\usage{
qLogLogistic(p, a, l)
}
\arguments{
\item{p}{Probability.}

\item{a}{Shape.}

\item{l}{Rate.}
}
\description{
Quantile function for the log-logistic distribution. See \code{\link{fit.LogLogistic}}
for the parameterization.
}
