% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{resample}
\alias{resample}
\title{Resample Data}
\usage{
resample(data)
}
\arguments{
\item{data}{Input data.farme}
}
\value{
Data.frame resampled from the input.
}
\description{
Generates a resampled data.frame with replacement.
}
