% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rGenGamma}
\alias{rGenGamma}
\title{Random Generation from the Generalized Gamma Distribution}
\usage{
rGenGamma(n, a = 1, b = 1, l = 1, p = 0)
}
\arguments{
\item{n}{Sample size.}

\item{a}{First shape parameter, \eqn{\alpha}.}

\item{b}{Second shape parameter, \eqn{\beta}. For the standard gamma
distribution, set \eqn{\beta=1}.}

\item{l}{Rate.}

\item{p}{Expected censoring proportion.}
}
\value{
A data.frame with two columns, the observation time and status
  indicator.
}
\description{
Generates generalized gamma random deviates with shape parameters 
\eqn{(\alpha,\beta)}, and rate paramter \eqn{\lambda}. See 
\code{\link{fit.GenGamma}} for the parameterization. If a censoring 
proportion \eqn{p} is provided, the deviates are subject to non-informative 
random right censoring.
}
\examples{
D = rGenGamma(n=1e3,a=2,b=2,l=2,p=0.2);
}
