% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rLogLogistic}
\alias{rLogLogistic}
\title{Random Generation from the Log-Logistic Distribution}
\usage{
rLogLogistic(n, a = 4, l = 1, p = 0)
}
\arguments{
\item{n}{Sample size.}

\item{a}{Shape.}

\item{l}{Rate.}

\item{p}{Expected censoring proportion.}
}
\value{
A data.frame with two columns, the observation time and status indicator.
}
\description{
Generates log-logistic random deviates with shape parameter \eqn{\alpha} and rate
paramter \eqn{\lambda}. See \code{\link{fit.LogLogistic}} for the parameterization. If
a censoring proportion \eqn{p} is provided, the deviates are subject to
non-informative random right censoring.
}
\examples{
D = rLogLogistic(n=1e3,a=4,l=1,p=0.2);
}
