% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{Rolling aggregate functions}
\alias{Rolling aggregate functions}
\alias{rolling.sum}
\title{Rolling aggregate functions}
\usage{
rolling.sum(x, w, type = "lag", na.rm = FALSE)
}
\arguments{
\item{x}{An atomic vector of class numeric.}

\item{w}{An integer value specifying the width of the windows over which to calculate the rolling aggregate.}

\item{type}{Either one of "lag" (window ends with current observation), "mid" (current observation is in the middle of the window) or "lead" (window starts with current observation).}

\item{na.rm}{A logical value indicating whether \code{NA} values should be ignored when calculating the rolling aggregates.}
}
\description{
Functions for calculating different rolling aggregates.
}
\details{
\itemize{
\item If \code{na.rm} is set to \code{TRUE} and there are no valid values inside a window, \code{rolling.sum} will return the value 0.
\item \code{rolling.var} returns the unbiased sample estimator \eqn{Var(x) = \frac{1}{n-1} \sum_i^n (x_i - \bar{x})^2}{Var(x) = (1 / (n - 1)) * sum_i (x_i - mu)^2}.
\item \code{rolling.sd} simply calls \code{rolling.var} and returnes the square root of the rolling variances.
\item The skewness measures the direction and degree of a distrubution's asymmetry.  A symmetric distribution has a skewness of zero, while a "left-skewed" (or "left-tailed") distribution (where the arithmetic mean is less than the median) has a negative skewness.  The Fisher-Pearson standardised moment coefficient, as discussed by Joanes and Gill (1998) is used to calculate the rolling skewness:
\deqn{\frac{\sqrt{n(n-1)}}{n(n-2)} \left( \frac{\sum_i^n (x_i-\bar{x})^3}{\left( \frac{1}{n} \sum_i^n (x_i-\bar{x})^2 \right)^{3/2}} \right)}{(sqrt(n(n - 1))/(n(n - 2)) ((sum_i (x_i - mu)^3)/((1 / n) sum_i (x_i - mu)^2)^(3 / 2))}
\item The kurtosis describes the "tailedness" of a distribution.  Using \eqn{s2 = \sum_i^n (x_i-\bar{x})^2}{s2 = sum_i (x_i - mu)^2}, \eqn{s4 = \sum_i^n (x_i-\bar{x})^4}{s4 = sum_i (x_i - mu)^4} and \eqn{Var(x) = \frac{s2}{n-1}}{Var(x) = s2 / (n - 1)}, the kurtosis is defined as follows:
\deqn{kurtosis = \frac{n(n-1)}{(n-1)(n-2)(n-3)} \left( \frac{s4}{Var(x)^2} \right) -3 \frac{(n-1)^2}{(n-2)(n-3)}.}{kurtosis = (n(n - 1) / ((n - 1)(n - 2)(n - 3))) (s4 / Var(x)^2) - 3(((n - 1)^2 / ((n - 2)(n - 3)))).}
The formula is can be found in Sheskin (2000) and yields an expected kurtosis of zero for a Gaussian distribution.
 }
}
\examples{
x <- rnorm(100)
rolling.sum(x = x, w = 10)

}
\references{
Sheskin, D.J. (2000) Handbook of Parametric and Nonparametric Statistical Procedures, Second Edition. Boca Raton, Florida: Chapman & Hall/CRC.

Joanes, D. N. and Gill, C. A. (1998), Comparing measures of sample skewness and kurtosis. Journal of the Royal Statistical Society: Series D (The Statistician), 47: 183-189. doi:10.1111/1467-9884.00122
}
