% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successions.R
\name{Successions}
\alias{Successions}
\alias{successions}
\title{Count successions}
\usage{
successions(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A list containing the indices of the successions (\code{index}), the total number of successions (\code{successions}), the unique value of each succession (\code{value}) and the lengths of the successions (\code{length})
}
\description{
Counts the lengths of successions of identical values in a vector.
}
\examples{
set.seed(7)
x <- sample(LETTERS[1:3], 10, replace=TRUE, prob = c(0.2, 0.2, 0.6))
x
successions(x)

}
