% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal-package.R
\docType{package}
\name{Temporal-package}
\alias{Temporal-package}
\alias{Temporal}
\alias{_PACKAGE}
\title{Temporal: Parametric Time to Event Analysis}
\description{
Maximum likelihood estimation and inference for parametric survival models
with support for right-censored data. FitParaSurv fits exponential, gamma,
generalized gamma, log-normal, and Weibull distributions. CompParaSurv
compares two treatment arms with respect to the mean, median, and restricted
mean survival time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zrmacc/Temporal}
  \item Report bugs at \url{https://github.com/zrmacc/Temporal/issues}
}

}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
