% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageHelp.R
\docType{package}
\name{Temporal}
\alias{Temporal}
\alias{Temporal-package}
\title{Temporal: Parametric Survival Analysis}
\description{
This package performs estimation and inference on parametric survival curves.
See \code{\link{fitParaSurv}} for maximum likelihood estimation of model parameters.
See \code{\link{compParaSurv}} for contrasting the survival experience of two 
treatment arms.
}
\author{
Zachary R. McCaw
}
