% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weibull.R
\name{WeiInit}
\alias{WeiInit}
\title{Weibull Initialization.}
\usage{
WeiInit(data, init)
}
\arguments{
\item{data}{Data.frame.}

\item{init}{Initialization list.}
}
\value{
Numeric initial value for shape.
}
\description{
Weibull Initialization.
}
