% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weibull.R
\name{WeiRate}
\alias{WeiRate}
\title{Weibull Rate MLE}
\usage{
WeiRate(data, shape)
}
\arguments{
\item{data}{Data.frame.}

\item{shape}{Shape parameter.}
}
\value{
Numeric rate.
}
\description{
Profile MLE of the Weibull rate as a function of the shape.
}
