% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weibull.R
\name{WeiScore}
\alias{WeiScore}
\title{Weibull Profile Score for Shape}
\usage{
WeiScore(data, shape)
}
\arguments{
\item{data}{Data.frame.}

\item{shape}{Shape parameter.}
}
\value{
Numeric score.
}
\description{
Profile score equation for the Weibull shape parameter.
}
