% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{CheckDist}
\alias{CheckDist}
\title{Check Distribution}
\usage{
CheckDist(dist)
}
\arguments{
\item{dist}{String, distribution name.}
}
\value{
None.
}
\description{
Check whether the distribution selected is available.
}
