% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{checkInit}
\alias{checkInit}
\title{Check Initialization}
\usage{
checkInit(dist, init)
}
\arguments{
\item{dist}{String, distribution name.}

\item{init}{Numeric vector, initialization.}
}
\value{
Logical indication of whether the initialization was valid.
}
\description{
Check whether the initialization is valid.
}
