% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{checkStatus}
\alias{checkStatus}
\title{Status Check}
\usage{
checkStatus(n, status)
}
\arguments{
\item{n}{Integer, sample size.}

\item{status}{0/1 status indicator.}
}
\value{
Logical indicator of whether the status indicator was properly
  formatted.
}
\description{
Function to ensure the status indicator is properly formatted
}
