% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{CheckTheta}
\alias{CheckTheta}
\title{Check Theta}
\usage{
CheckTheta(dist, theta)
}
\arguments{
\item{dist}{String, distribution.}

\item{theta}{Numeric, parameter vector.}
}
\value{
None.
}
\description{
Function to check the appropriate number of parameters are supplied for the
selected distribution. Used by \code{\link{GenData}}.
}
