% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{contrast-class}
\alias{contrast-class}
\title{Contrast of Survival Distributions.}
\description{
Defines the object class returned by the comparison function.
}
\section{Slots}{

\describe{
\item{\code{Dist1}}{Distribution fit to the target group.}

\item{\code{Dist0}}{Distribution fit to the reference group.}

\item{\code{Model1}}{Fitted model for the target group.}

\item{\code{Model0}}{Fitted model for the reference group.}

\item{\code{Contrast}}{Model contrasts.}
}}

