% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{methods}
\name{show,contrast-method}
\alias{show,contrast-method}
\title{Show for Contrasts}
\usage{
\S4method{show}{contrast}(object)
}
\arguments{
\item{object}{A \code{contrast} object.}
}
\description{
Show for Contrasts
}
