% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{defaultParam}
\alias{defaultParam}
\title{Set Default Parameters}
\usage{
defaultParam(dist)
}
\arguments{
\item{dist}{String, distribution name.}
}
\value{
Numeric parameter vector
}
\description{
Function to select default parameter values for each distribution.
}
