% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncGamma.R
\name{dIncGamma}
\alias{dIncGamma}
\title{Derivatives of the Incomplete Gamma Function}
\usage{
dIncGamma(a, s, Dirn = "a", Order = 1)
}
\arguments{
\item{a}{Value of \eqn{\alpha} at which to evaluate.}

\item{s}{Value of \eqn{s} at which to evaluate.}

\item{Dirn}{Direction in which to differentiate. Select from among "a", "s",
and "as".}

\item{Order}{Order of the derivative, if the direction is either "a", or "s".
Select from among 1 and 2.}
}
\description{
Evaluates derivatives of the upper incomplete gamma function, defined as:
\deqn{\Gamma(\alpha,s) = \int_{s}^{\infty}u^{\alpha-1}e^{-u}du}
}
