% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{estRatio}
\alias{estRatio}
\title{Ratio of Estimates}
\usage{
estRatio(t1, s1, t0, s0, sig = 0.05)
}
\arguments{
\item{t1}{Treatment estimate}

\item{s1}{Treatment standard error}

\item{t0}{Reference estimate}

\item{s0}{Reference standard error}

\item{sig}{Significance level}
}
\description{
Calculate CIs and p-value for the ratio of estimated parameters
}
