% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{fit-class}
\alias{fit-class}
\title{Fitted Survival Distribution}
\description{
Defines the object class returned by fitting functions.
}
\section{Slots}{

\describe{
\item{\code{Distribution}}{Fitted distribution, string.}

\item{\code{Parameters}}{Parameters, data.frame.}

\item{\code{Information}}{Information components, matrix.}

\item{\code{Outcome}}{Properties of the fitted distribution, data.frame.}

\item{\code{RMST}}{Estimated restricted mean survival times, data.frame}

\item{\code{S}}{Fitted survival function, function.}
}}

