% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{methods}
\name{show,fit-method}
\alias{show,fit-method}
\title{Show Method for Fitted Survival Distributions}
\usage{
\S4method{show}{fit}(object)
}
\arguments{
\item{object}{An object of class \code{fit}.}
}
\description{
Show Method for Fitted Survival Distributions
}
