% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{fit.LogNormal}
\alias{fit.LogNormal}
\title{Log-Normal Parameter Estimation}
\usage{
fit.LogNormal(time, status, alpha = 0.05, eps = 1e-06, maxit = 10)
}
\arguments{
\item{time}{Observation times.}

\item{status}{Status indicator, coded as 1 if an event was observed, 0 if censored.}

\item{alpha}{Significance level, for CIs.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}
}
\value{
An object of class \code{fit} containing the following:
\describe{
 \item{Parameters}{The estimated location \eqn{\mu} and scale \eqn{\sigma}.}
 \item{Information}{The observed information matrix.}
 \item{Outcome}{The fitted mean, median, and variance.}
}
}
\description{
Estimates parameters for log-normal event times subject to non-informative
right censoring. The log-normal distribution is parameterized in terms
of the location \eqn{\mu} and scale \eqn{\sigma}:
\deqn{f(t) = \phi\left(\frac{\ln t-\mu}{\sigma}\right)\frac{1}{t\sigma}, t>0}
}
\details{
In the log-normal density, \eqn{\phi(\cdot)} denotes the standard normal pdf.
}
