% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{fitParaSurv}
\alias{fitParaSurv}
\title{Fit Parametric Survival Distribution}
\usage{
fitParaSurv(time, status, dist = "weibull", alpha = 0.05, eps = 1e-06,
  maxit = 10)
}
\arguments{
\item{time}{Observation times.}

\item{status}{Status indicator, coded as 1 if an event was observed, 0 if
censored.}

\item{dist}{Distribution to fit, selected from among: exp, gamma,
log-logistic, log-normal, and weibull.}

\item{alpha}{Significance level, for CIs.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}
}
\value{
An object of class \code{fit} containing the following: \describe{ 
  \item{Parameters}{The estimated rate \eqn{\lambda}.} \item{Information}{The
  observed information.} \item{Outcome}{The fitted mean, median, and
  variance.} }
}
\description{
Estimates parameters for parametric event times subject to non-informative 
right censoring. Available distributions include: exponential, gamma, 
log-logistic, log-normal, and Weibull.
}
\examples{
# Generate cenored gamma data
D = rGamma(n=1e3,a=2,l=2,p=0.2);
# Fit gamma distribution
M = fitParaSurv(time=D$time,status=D$status,dist="gamma");

# Generate cenored weibull data
D = rWeibull(n=1e3,a=2,l=2,p=0.2);
# Fit weibull distribution
M = fitParaSurv(time=D$time,status=D$status,dist="weibull");
}
