% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{permP}
\alias{permP}
\title{Permutation P Value}
\usage{
permP(d1, d0, tau, dist1, dist0, B = 1999)
}
\arguments{
\item{d1}{Target data.frame containing time and status.}

\item{d0}{Reference data.frame containing time and status.}

\item{tau}{Optional truncation times for calculating restricted mean survival
time.}

\item{dist1}{String, target distribution.}

\item{dist0}{String, reference distribution.}

\item{B}{Number of resamples.}
}
\value{
List containing data.frames with the bootstrap CIs for the location
  and RMST estimates.
}
\description{
Calculates permutation p-values for location and RMST estimates.
}
