% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.contrast}
\alias{print.contrast}
\title{Print Method for a Contrast of Survival Distributions.}
\usage{
\method{print}{contrast}(x, ...)
}
\arguments{
\item{x}{A \code{contrast} object.}

\item{...}{Unused.}
}
\description{
Print method for an object of class \code{contrast}.
}
