% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.fit}
\alias{print.fit}
\title{Print for Fitted Survival Models}
\usage{
\method{print}{fit}(x, ...)
}
\arguments{
\item{x}{A \code{fit} object.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{fit}.
}
