% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{qLogLogistic}
\alias{qLogLogistic}
\title{Quantile Function for the Log-Logistic Distribution}
\usage{
qLogLogistic(p, a = 1, l = 1)
}
\arguments{
\item{p}{Probability.}

\item{a}{Shape.}

\item{l}{Rate.}
}
\value{
Scalar quantile.
}
\description{
Quantile function for the log-logistic distribution. See \code{\link{fit.LogLogistic}}
for the parameterization.
}
\examples{
# Median of standard log-logistic distribution
qLogLogistic(p=0.5);
}
