% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{qWeibull}
\alias{qWeibull}
\title{Quantile Function for the Weibull Distribution}
\usage{
qWeibull(p, a, l)
}
\arguments{
\item{p}{Probability.}

\item{a}{Shape.}

\item{l}{Rate.}
}
\description{
Quantile function for the Weibull distribution. See \code{\link{fit.Weibull}}
for the parameterization.
}
