% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rGamma}
\alias{rGamma}
\title{Simulation from the Gamma Distribution}
\usage{
rGamma(n, a = 1, l = 1, p = 0)
}
\arguments{
\item{n}{Sample size.}

\item{a}{Shape.}

\item{l}{Rate.}

\item{p}{Expected censoring proportion.}
}
\value{
Data.frame including the observation times and status.
}
\description{
Generates gamma event times with shape parameter \eqn{\alpha} and rate
parameter \eqn{\lambda}. See \code{\link{FitGamma}} for the parameterization. If
a censoring proportion \eqn{p} is provided, the event times are subject to
non-informative random right censoring.
}
