% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rGamma}
\alias{rGamma}
\title{Random Generation from the Gamma Distribution}
\usage{
rGamma(n, a = 1, l = 1, p = 0)
}
\arguments{
\item{n}{Sample size.}

\item{a}{Shape.}

\item{l}{Rate.}

\item{p}{Expected censoring proportion.}
}
\value{
A data.frame with two columns, the observation time and status indicator.
}
\description{
Generates gamma random deviates with shape parameter \eqn{\alpha} and rate
paramter \eqn{\lambda}. See \code{\link{fit.Gamma}} for the parameterization. If
a censoring proportion \eqn{p} is provided, the deviates are subject to
non-informative random right censoring.
}
\examples{
D = rGamma(n=1e3,a=2,l=2,p=0.2);
}
