% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rLogNormal}
\alias{rLogNormal}
\title{Random Generation from the Log-Normal Distribution}
\usage{
rLogNormal(n, m = 0, s = 1, p = 0)
}
\arguments{
\item{n}{Sample size.}

\item{m}{Location.}

\item{s}{Scale.}

\item{p}{Expected censoring proportion.}
}
\value{
A data.frame with two columns, the observation time and status indicator.
}
\description{
Generates log-normal random deviates with location parameter \eqn{\mu} and scale
paramter \eqn{\sigma}. See \code{\link{fit.LogNormal}} for the parameterization. If
a censoring proportion \eqn{p} is provided, the deviates are subject to
non-informative random right censoring.
}
\examples{
D = rLogNormal(n=1e3,m=0,s=1,p=0.2);
}
