% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rWeibull}
\alias{rWeibull}
\title{Random Generation from the Weibull Distribution}
\usage{
rWeibull(n, a = 1, l = 1, p = 0)
}
\arguments{
\item{n}{Sample size.}

\item{a}{Shape.}

\item{l}{Rate.}

\item{p}{Expected censoring proportion.}
}
\value{
A data.frame with two columns, the observation time and status indicator.
}
\description{
Generates Weibull random deviates with shape parameter \eqn{\alpha} and rate
paramter \eqn{\lambda}. See \code{\link{fit.Weibull}} for the parameterization. If
a censoring proportion \eqn{p} is provided, the deviates are subject to
non-informative random right censoring.
}
\examples{
D = rWeibull(n=1e3,a=2,l=2,p=0.2);
}
