% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal-package.R
\docType{package}
\name{Temporal-package}
\alias{Temporal}
\alias{Temporal-package}
\title{Temporal: Parametric Time to Event Analysis}
\description{
Performs maximum likelihood based estimation and inference on time to event data, possibly subject to non-informative right censoring. FitParaSurv() provides maximum likelihood estimates of model parameters and distributional characteristics, including the mean, median, variance, and restricted mean. CompParaSurv() compares the mean, median, and restricted mean survival experiences of two treatment groups. Candidate distributions include the exponential, gamma, generalized gamma, log-normal, and Weibull.
}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
